+++ tests/test.lua
@@ -1,6 +1,14 @@
 package.path = "../?/init.lua;../?.lua;"..package.path
-multi, thread = require("multi"):init{print=true,warn=true,error=true}
+multi, thread = require("multi"):init{print=true,warn=true,error=true,debugging=true}
 require("multi.integration.priorityManager")
+
+multi.debugging.OnObjectCreated(function(obj, process)
+	multi.print("Created:", obj.Type, "in", process.Type, process:getFullName())
+end)
+
+multi.debugging.OnObjectDestroyed(function(obj, process)
+	multi.print("Destroyed:", obj.Type, "in", process.Type, process:getFullName())
+end)
 
 -- test = multi:newProcessor("Test")
 -- test:setPriorityScheme(multi.priorityScheme.TimeBased)
@@ -66,6 +74,11 @@ proc:newThread(function()
 		multi.warn("Everything is a thread in this proc!")
 		thread.sleep(1)
 	end
+end)
+
+proc:newAlarm(5):OnRing(function(a)
+	multi.print(";) Goodbye")
+	a:Destroy()
 end)
 
 multi:mainloop()